Fix clickability of IDs in the list of segment values
The latte template expected to have array keys keyed by column
names, but because of array_values it was indexed by numbers.
Because of that, template couldn't decide if the column should be
clickable or not.

This only fixes clickability of IDs for users segments that
was already implemented. It doesn't add clickability for other
tables.

remp/crm#2943